# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Data Karyawan Service API",
        "description": "REST API untuk manajemen data karyawan",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get all employees",
                "operationId": "1c7422337e6554c16e1d0e91f994fd58",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create employee",
                "operationId": "3ff8e49d36a48e1e830416022f8c2bc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nip": {
                                        "type": "string"
                                    },
                                    "nama": {
                                        "type": "string"
                                    },
                                    "jabatan": {
                                        "type": "string"
                                    },
                                    "departemen": {
                                        "type": "string"
                                    },
                                    "gaji_pokok": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee by ID",
                "operationId": "635dcc60d9cc155661fd3b0de13187e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/messages/publish": {
            "post": {
                "tags": [
                    "RabbitMQ"
                ],
                "summary": "Publish event ke RabbitMQ",
                "operationId": "a445713ced5e35d2f30845ec5c1da961",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "event_type": {
                                        "type": "string",
                                        "example": "EmployeeCreated"
                                    },
                                    "payload": {
                                        "type": "string",
                                        "example": "{\"employee_id\":1}"
                                    },
                                    "bearer_token": {
                                        "type": "string",
                                        "example": "eyJ..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event published"
                    },
                    "500": {
                        "description": "Publish error"
                    }
                }
            }
        },
        "/api/v1/audit/send": {
            "post": {
                "tags": [
                    "SOAP Audit"
                ],
                "summary": "Kirim audit ke SOAP server",
                "operationId": "eb5968c6d2de02649b9c45cb912c3be8",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "activity_name": {
                                        "type": "string",
                                        "example": "EmployeeDataAccess"
                                    },
                                    "transaction_data": {
                                        "type": "string",
                                        "example": "{\"action\":\"GET\",\"resource\":\"employees\"}"
                                    },
                                    "bearer_token": {
                                        "type": "string",
                                        "example": "eyJ..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Audit sent"
                    },
                    "500": {
                        "description": "SOAP error"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "SSO"
                ],
                "summary": "Login user via SSO",
                "operationId": "1ce8f7a861d8cb7183db0c6e858a16d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "warga30@ktp.iae.id"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "KtpDigital2026!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/auth/login-m2m": {
            "post": {
                "tags": [
                    "SSO"
                ],
                "summary": "Login M2M via API Key",
                "operationId": "e77aa85eb730f783342b491d55a9d005",
                "responses": {
                    "200": {
                        "description": "Token retrieved"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/auth/health": {
            "get": {
                "tags": [
                    "SSO"
                ],
                "summary": "SSO Health Check",
                "operationId": "a297de04ebe92473a6ed85aabcf2fd14",
                "responses": {
                    "200": {
                        "description": "SSO is up"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Employees",
            "description": "Employees"
        },
        {
            "name": "RabbitMQ",
            "description": "RabbitMQ"
        },
        {
            "name": "SOAP Audit",
            "description": "SOAP Audit"
        },
        {
            "name": "SSO",
            "description": "SSO"
        }
    ]
}